[day24-2] Check that I didn't get lucky...
It's unlikely, but I never actually pruned the heuristic-pruned result
states to just z=0.
My input might be lucky in that it has a b<0 last, but also my hunch is
that that's basically required. Since I do-or-die on choosing input/w
such that z=0 when b<0 (aka a==26), all my valid model numbers are
already z=0.